Fix an error in the equations of motion. Allow w != kappa
New equations of motion and initial conditions.
There was previously a term K tau'^2 missing from
the lambda'' -equation. The term is proportional to
n^2 tau'^2, so the error is expected to be small,
and to not change the results in 1312.5199 significantly.
Preliminarily, the numerics seem to also support such
a conclusion.

The new equations of motion also implement w != kappa,
allowing a wider range of potentials.