Add a check on Veff in tauhFromQuarkMass
Add a check that Veff > 0 before attempting anything
else in tauhFromQuarkMass. This has a very minor
negative effect on runtime when the check is
passed, and a significant positive effect when not.

In addition, this is necessary in Mathematica 9,
which would attempt to evaluate bignums for very
large values of tauh, effectively choking the
entire computation.