Enforce order of sum in NIntegralFunction
Add the possibility, on by default, to enforce the
order of the sum to be from the lower limit to the
upper limit, even when their numerical order is
reversed, when using the IntegrationGrid -method
in NIntegralFunction. This allows some integrals
to remain numerically  stable even when they go
through several orders of magnitude change.