Return correct fscale when fpInitialValue != 1
Previously if fpInitialValue was not 1, the thermo
extraction functions would give invalid results,
as they assume f' was originally one, such that the
physical value is given by fscale^2. Fix this by 
returning fscale * Sqrt[f'(A_h)] as fscale, which
is the number by which the f'[A_h] = 1 solution would
have to be scaled, independent of whether this was
the initial condition actually used.

This allows adjusting the initial condition for better
numerical stability.